requirements.txt
llama-cpp-python
colorama
pyfiglet
